Unclutter LibUsb type; extract device enumeration
methods to static LibUsbDeviceEnum helper type
