+++ src/LibUsbSharp/LibUsb.cs
@@ -2,7 +2,6 @@ using System.Collections.Concurrent;
 using System.Runtime.InteropServices;
 using LibUsbSharp.Descriptor;
 using LibUsbSharp.Internal;
-using LibUsbSharp.Internal.Descriptor;
 using LibUsbSharp.Internal.Hotplug;
 using Microsoft.Extensions.Logging;
 using Microsoft.Extensions.Logging.Abstractions;
@@ -159,7 +158,7 @@ public sealed class LibUsb : ILibUsb
     )
     {
         // TODO: Test on macOS and linux; "most functions that take a device handle are not safe"
-        var result = TryGetDeviceDescriptor(device, out var deviceDescriptor);
+        var result = LibUsbDeviceEnum.TryGetDeviceDescriptor(device, out var deviceDescriptor);
         if (result != LibUsbResult.Success)
         {
             _logger.LogWarning(
@@ -188,24 +187,12 @@ public sealed class LibUsb : ILibUsb
         lock (_lock)
         {
             CheckDisposed();
-            var result = libusb_get_device_list(GetInitializedContextOrThrow(), out var listPtr);
-            try
-            {
-                return result >= 0
-                    ? GetDeviceDescriptors(listPtr)
-                        .Select(d => d.Descriptor)
-                        .Where(d =>
-                            (vendorId is null || vendorId == d.VendorId)
-                            && (productIds is null || productIds.Contains(d.ProductId))
-                        )
-                        .Cast<IUsbDeviceDescriptor>()
-                        .ToList()
-                    : throw LibUsbException.FromError(result, "Failed to get device list.");
-            }
-            finally
-            {
-                libusb_free_device_list(listPtr, 1);
-            }
+            return LibUsbDeviceEnum.GetDeviceList(
+                _logger,
+                GetInitializedContextOrThrow(),
+                vendorId,
+                productIds
+            );
         }
     }
 
@@ -269,7 +256,8 @@ public sealed class LibUsb : ILibUsb
 
     private UsbDevice OpenListDeviceUnlocked(nint context, nint listPtr, string deviceKey)
     {
-        var (descriptorPtr, descriptor) = GetDeviceDescriptors(listPtr)
+        var (descriptorPtr, descriptor) = LibUsbDeviceEnum
+            .GetDeviceDescriptors(_logger, listPtr)
             .FirstOrDefault(d => d.Descriptor.DeviceKey == deviceKey);
         if (descriptorPtr == IntPtr.Zero)
         {
@@ -278,7 +266,10 @@ public sealed class LibUsb : ILibUsb
                 "Failed to get device from list."
             );
         }
-        var configResult = TryGetConfigDescriptor(descriptorPtr, out var configDescriptor);
+        var configResult = LibUsbDeviceEnum.TryGetConfigDescriptor(
+            descriptorPtr,
+            out var configDescriptor
+        );
         if (configResult != 0)
         {
             throw LibUsbException.FromResult(
@@ -316,86 +307,6 @@ public sealed class LibUsb : ILibUsb
             }
             libusb_close(handle);
         }
-    }
-
-    /// <summary>
-    /// Get cached USB device descriptors for a given, alrady in memory, device descriptor list.
-    /// </summary>
-    /// <param name="listPtr">Pointer to device list returned by libusb_get_device_list.</param>
-    private IEnumerable<(nint DescriptorPtr, UsbDeviceDescriptor Descriptor)> GetDeviceDescriptors(
-        nint listPtr
-    )
-    {
-        var offset = 0;
-        nint descriptorPtr;
-        while ((descriptorPtr = Marshal.ReadIntPtr(listPtr, offset * IntPtr.Size)) != IntPtr.Zero)
-        {
-            var result = TryGetDeviceDescriptor(descriptorPtr, out var descriptor);
-            if (result != LibUsbResult.Success)
-            {
-                _logger.LogWarning(
-                    "{LibUsb} get device descriptor failed. {ErrorMessage}",
-                    LibraryName,
-                    result.GetMessage()
-                );
-            }
-            else if (descriptor!.Value.BcdUsb > 0)
-            {
-                yield return (descriptorPtr, descriptor!.Value);
-            }
-            offset++;
-        }
-    }
-
-    /// <summary>
-    /// Get the cached USB device descriptor for a given, alrady in memory, device descriptor list.
-    /// NOTE: since libusb-1.0.16, LIBUSBX_API_VERSION >= 0x01000102, this function always succeeds.
-    /// </summary>
-    private static LibUsbResult TryGetDeviceDescriptor(
-        nint deviceDescriptorPtr,
-        out UsbDeviceDescriptor? descriptor
-    )
-    {
-        descriptor = null;
-        var result = libusb_get_device_descriptor(deviceDescriptorPtr, out var partialDescriptor);
-        if (result == 0)
-        {
-            descriptor = new UsbDeviceDescriptor(
-                partialDescriptor,
-                libusb_get_bus_number(deviceDescriptorPtr),
-                libusb_get_device_address(deviceDescriptorPtr),
-                libusb_get_port_number(deviceDescriptorPtr)
-            );
-        }
-        return (LibUsbResult)result;
-    }
-
-    /// <summary>
-    /// Get the USB configuration descriptor for the currently active device configuration. This
-    /// is a non-blocking function which does not involve any requests being sent to the device.
-    /// </summary>
-    private static LibUsbResult TryGetConfigDescriptor(
-        nint deviceDescriptorPtr,
-        out IUsbConfigDescriptor? descriptor
-    )
-    {
-        descriptor = null;
-        var result = libusb_get_active_config_descriptor(deviceDescriptorPtr, out var configPtr);
-        if (result != 0)
-        {
-            return (LibUsbResult)result;
-        }
-        try
-        {
-            descriptor = Marshal
-                .PtrToStructure<LibUsbConfigDescriptor>(configPtr)
-                .ToUsbInterfaceDescriptor();
-        }
-        finally
-        {
-            libusb_free_config_descriptor(configPtr);
-        }
-        return LibUsbResult.Success;
     }
 
     /// <summary>
@@ -525,53 +436,6 @@ public sealed class LibUsb : ILibUsb
 
     [DllImport(LibraryName, CallingConvention = CallingConvention.Cdecl)]
     private static extern void libusb_free_device_list(IntPtr listPtr, int unrefDevices);
-
-    [DllImport(LibraryName, CallingConvention = CallingConvention.Cdecl)]
-    private static extern int libusb_get_device_descriptor(
-        IntPtr deviceDescriptorPtr,
-        out LibUsbDeviceDescriptor deviceDescriptor
-    );
-
-    /// <summary>
-    /// Get the number of the bus that a device is connected to.
-    /// </summary>
-    [DllImport(LibraryName, CallingConvention = CallingConvention.Cdecl)]
-    private static extern byte libusb_get_bus_number(IntPtr deviceDescriptorPtr);
-
-    /// <summary>
-    /// Get the address of the device on the bus it is connected to.
-    /// </summary>
-    [DllImport(LibraryName, CallingConvention = CallingConvention.Cdecl)]
-    private static extern byte libusb_get_device_address(IntPtr deviceDescriptorPtr);
-
-    /// <summary>
-    /// Get the number of the port that a device is connected to.
-    ///
-    /// The number returned by this call is usually guaranteed to be uniquely tied to a physical
-    /// port, meaning that different devices plugged on the same physical port should return the
-    /// same port number. But there is no guarantee that the port number returned by this call will
-    /// remain the same, or even match the order in which ports are numbered on the HUB/HCD.
-    /// </summary>
-    [DllImport(LibraryName, CallingConvention = CallingConvention.Cdecl)]
-    private static extern byte libusb_get_port_number(IntPtr deviceDescriptorPtr);
-
-    /// <summary>
-    /// Get the USB configuration descriptor for the currently active configuration. This is
-    /// a non-blocking function which does not involve any requests being sent to the device.
-    /// </summary>
-    [DllImport(LibraryName, CallingConvention = CallingConvention.Cdecl)]
-    private static extern int libusb_get_active_config_descriptor(
-        IntPtr deviceDescriptorPtr,
-        out IntPtr deviceConfigPtr
-    );
-
-    /// <summary>
-    /// Free a configuration descriptor obtained from
-    /// libusb_get_active_config_descriptor() or libusb_get_config_descriptor()
-    /// </summary>
-    /// <param name="configPtr"></param>
-    [DllImport(LibraryName, CallingConvention = CallingConvention.Cdecl)]
-    private static extern void libusb_free_config_descriptor(IntPtr configPtr);
 
     /// <summary>
     /// Open a device and obtain a device handle. Allows you to perform I/O on the device.
